+++ Dockerfile
@@ -1,4 +1,4 @@
-  FROM lukemathwalker/cargo-chef:0.1.50-rust-buster AS chef
+FROM lukemathwalker/cargo-chef:0.1.50-rust-buster AS chef
 WORKDIR /app
 
 FROM chef AS planner
@@ -22,12 +22,12 @@ FROM debian:bullseye-slim as base
 WORKDIR /app
 RUN apt-get update -y && \
   apt-get install -y --no-install-recommends \
-    ca-certificates \
-    libpq5 \
-    libssl1.1 \
+  ca-certificates \
+  libpq5 \
+  libssl1.1 \
   && \
   rm -rf /var/lib/apt/lists/*
 
 COPY --from=builder /app/target/release/holaplex-hub-identities /usr/local/bin
-ENTRYPOINT [ "/usr/local/bin/holaplex-hub-identities" ]
+ENTRYPOINT ["/usr/local/bin/holaplex-hub-identities"]
