+++ api/src/graphql/mod.rs
@@ -1,5 +1,5 @@
 pub mod dataloader;
 pub mod mutation;
-pub mod object;
+pub mod objects;
 pub mod query;
 pub mod schema;
