+++ api/src/graphql/objects/mod.rs
@@ -2,8 +2,10 @@ mod credential;
 mod member;
 mod owner;
 mod user;
+mod webhook;
 
 pub use credential::Credential;
 pub use member::Member;
 pub use owner::Owner;
 pub use user::User;
+pub use webhook::Webhook;
