+++ api/src/graphql/query/mod.rs
@@ -2,8 +2,15 @@ mod credential;
 mod member;
 mod owner;
 mod user;
+mod webhook;
 
 // Add your other ones here to create a unified Query object
 // e.x. Query(SomeQuery, OtherQuery, OtherOtherQuery)
 #[derive(async_graphql::MergedObject, Default)]
-pub struct Query(user::Query, owner::Query, member::Query, credential::Query);
+pub struct Query(
+    user::Query,
+    owner::Query,
+    member::Query,
+    credential::Query,
+    webhook::Query,
+);
