+++ api/src/graphql/query/webhook.rs
@@ -1,19 +1,19 @@
 use async_graphql::{Context, Object, Result};
 use hub_core::uuid::Uuid;
 
-use crate::graphql::objects::Credential;
+use crate::graphql::objects::Webhook;
 
 #[derive(Default)]
 pub struct Query;
 
-#[Object(name = "CredentialQuery")]
+#[Object(name = "WebhookQuery")]
 impl Query {
     #[graphql(entity)]
-    async fn find_credential_by_created_by_id(
+    async fn find_webhook_by_created_by_id(
         &self,
         _ctx: &Context<'_>,
         #[graphql(key)] created_by_id: Uuid,
-    ) -> Result<Credential> {
-        Ok(Credential { created_by_id })
+    ) -> Result<Webhook> {
+        Ok(Webhook { created_by_id })
     }
 }
